requirements.txt
streamlit
pandas
yfinance
plotly
google-genai